requirements.txt
aiohttp==1.0.5
async-timeout==3.0.0
beautifulsoup4==4.6.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
discord.py==0.16.12
idna==2.7
multidict==4.3.1
pycparser==2.18
PyNaCl==1.0.1
requests==2.19.1
six==1.11.0
spotipy==2.4.4
urllib3==1.23
websockets==3.4
youtube-dl==2018.7.10
